refactor: modify createBridge function signature in Bridge module
